bsdsocket: canonicalise ioctl commands at K&R call boundaries
On riscv64 a compiled callee compares an int parameter as a full
register against sign-extended constants, trusting the ABI's canonical
form. in_control() was reached through unprototyped calls that passed
the command zero-extended, and rtioctl() received it zero-extended
across the library-call boundary, where a plain cast is elided because
the compiler may assume a parameter is already canonical - hence the
volatile round-trip. Every socket-level interface and route ioctl came
back EOPNOTSUPP or EINVAL, so dhclient could never configure an
interface. x86-64 compiles the same comparisons 32 bits wide and never
noticed; udp6_usrreq.c had already been given the equivalent correction
on the IPv6 side.

The now-visible prototype also exposed a stray second argument to the
one-argument in_broadcast(); dropped.

+++ workbench/network/stacks/AROSTCP/bsdsocket/netinet/tcp_usrreq.c
@@ -63,6 +63,7 @@
 #include <netinet/in_var.h>
 #include <netinet/ip_var.h>
 #include <netinet/tcp.h>
+#include <netinet/in_protos.h>
 #include <netinet/tcp_fsm.h>
 #include <netinet/tcp_seq.h>
 #include <netinet/tcp_timer.h>
@@ -114,7 +115,7 @@ struct mbuf *m, *nam, *control;
 #endif
 
     if(req == PRU_CONTROL)
-        return (in_control(so, (long)m, (caddr_t)nam,
+        return (in_control(so, (int)(long)m, (caddr_t)nam,
                            (struct ifnet *)control));
     if(control && control->m_len) {
         m_freem(control);
